mako: msm: acpuclock: Convert acpuclock drivers into platform drivers
msm: acpuclock: Convert acpuclock drivers into platform drivers
commit	bf430ebf4f7f05609d75ffbb9f562e7f2d0c4eb2

This allows the acpuclock drivers to initialize in a manner consistent
with other platform drivers used on MSM chipsets.

For targets without specific probe order dependencies involving
acpuclock, platform_device_register() is called during init_machine()
in the target's board file. platform_driver_probe() is then called
in a 'device' initcall. This is the case for:
  acpuclock-8x60, acpuclock-8960, acpuclock-9615, acpuclock-fsm9xxx

For the other targets, platform_driver_register() is called in a
postcore initcall and the driver probes as a result of calling
platform_device_register() during init_machine(). This is required
for the following drivers:
  acpuclock-7627, acpuclock-7x30, acpuclock-8x50

Specifically, these three drivers are used on targets where the CPUs
may be running from disableable clock sources that are shared with
other peripheral clocks. We must make sure that acpuclock has a
chance to assert a vote for the clock source that the CPU is currently
running from before other drivers using that same source have a chance
to disable it.

Change-Id: Iaa2544d705862c09c7e16b97cca86e817d766424