msm: bam_dmux: update SSR handling to reflect disconnect ack on by default
When subsystem restart of the modem is triggered, the SSR handling function
disables the disconnect ack feature.  Later, when the modem reboots, it
expects the disconnect ack feature to be enabled by default as it was when
the modem first booted.  This causes a problem when the modem times out
waiting for the disconnect ack and asserts.

Fix the issue by changing the SSR code in bam_dmux to reenable the
disconnect ack feature instead of disabling it.

(cherry picked from commit f62029d1c902342470816ad4920b08ea805a52d6)

Change-Id: I639695939f76a97ecc4d5d2bf44f05cb18fb1e41
CRs-Fixed: 377493
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>