Merge prima driver releases 3.2.0.1 and 3.2.0.2 from codeaurora.org
git clone git://codeaurora.org/external/wlan/prima.git
	d13512 prima: WLAN Driver Release 3.2.0.1
	43971f5 prima: WLAN Driver Release 3.2.0.2

prima: WLAN Driver Release 3.2.0.1

Integrate the following CLs from Perforce:

635258: Integrate the latest Riva API

635360: Cleaning up WDI_SendMsg for WDA requests with req callbacks
1. For WDI requests which have registered a request callback, inform
the WDA caller of the same via setting the return value (wdiStatus) to
WDI_STATUS_PENDING. This makes sure that WDA doesnt end up repeating
the functonality in the req callback for the WDI_STATUS_E_FAILURE case
2. General cleanup/restructuring of WDI_SendMsg.
(CR 356091)

635425: Add support for multicast MAC address list
Provision to configure multicast MAC address list "gMCAddrListEnable"
flag and support for Multicast MAC address List.  Currently Multicast
MAC address List is not supported. Nowit is supported in Riva and Host
changes are verified. In addition to this, one ini flag
"gMCAddrListEnable" is introduced to enable this support.
  # to enable Multicast MAC address list
  # Flags to enable filter Mcast MAC address list separately
  # gMCAddrListEnable=1 to enable the support
  # gMCAddrListEnable=0 to disable the support
  gMCAddrListEnable=1
Default value 0 if not configured in ini file.  (CR 371401)

635520: Fix a possible NULL pointer dereference in SME 'get statistics'
Fix a possible NULL pointer dereference while SME processes a 'get
statistics' de-register request. (CR 368659)

636116: Support multicast MAC address list

636138 & 636161: Remove hard tabs

636159: Avoid duplicate entries in gLimMgmtFrameRegistratinQueue
__limProcessSmeRegisterMgmtFrameReq is adding entry even there is an
entry for requested mgmt frame already exists.  This is creating
problem when wlan_hdd_request_remain_on_channel function get called
back to back.  Fix- First checking whether the entry for requested
mgmt frame exists in the list; if exists then deleting the entry and
adding new entry to make sure that there wont be no duplicate entries
in the list.

636356: Avoid unwanted reset of country code after disassoc/assoc failure
After dis-association or association failure, reset the country code
only if 11d is disabled and current operating country code and default
country codes are different.  If we reset the country code, driver is
flushing the scan results and which is triggering full scan. Full scan
is taking 3-4 sec and this delay is causing connection issues in
roaming.

636724: Fix for Prefetch Abort in Vos MC Thread
Fix to not invoke WDI callback function if it is NULL (CR 370598)

prima: WLAN Driver Release 3.2.0.2

Integrate the following CLs from Perforce:

636735: Insure PMK candidate notify event only sent up to supplicant for RSN
Insure that PMK candidate notify even only sent up to supplicant for
RSN associations (CR 362782)

636761,
636770,
636789: Feature for supporting roaming to an AP with better RSSI
This feature takes care of the station to roam to a new AP only if the
roamable/new AP has a better RSSI than current AP.  A new ini item has
been added named RoamRssiDiff using which we can configure how better
in terms of Rssi, the new AP that we want to roam should be. On
providing a value of Zero, the feature is disabled and the check for
an AP with better RSSI does not happen while roaming. (CR 366355)

636834: Set rssi filter period to FT value
if CCX is disable and LFR is enabled, rssi filter period was set to 20
(default value) because of which RSSI monitoring was not effective,
changes added to set rssi filter period to FT rssi value (5) even for
LFR enabled case also. (CR 372303)

636843: isFastRoamingEnabled passed during reassoc (CR 372303)

636970: Implemented getwlandbg for SoftAP
SoftAP does not have an ioctl to print the log levels and log
info. Implemented getwlandbg for SoftAP.  (CR 374467)

636989 & 636992: Remove hard tabs and DOS line endings

637100: Fix of WDA Stop timeout kernel panic issue.
WDA stop timeout should never happen. Increasing WDA stop timeout
after considering 1 pending request and cleaning up WDI pending queues
once driver unload is trigerred.  (CR 367261)

637246: Fixed WLAN_BAPReset() to avoid sending Disassoc on STA interface.
Due to some reason(Seen just after driver load and whenever
hdd_netdev_notifier_call get called with netdevice state
NETDEV_GOING_DOWN where we call WLANBAP_DeregisterFromHCI())
WLAN_BAPReset was getting called through BSL_Write -> bslWriteFinish.
Even If the BT-AMP session is not opened(in this case
btampHandle->sessionId will be zero(default value) and if
WLAN_BAPReset is called, BAP was issueing sme_RoamDisconnect with
sessionId 0 which is station interface which results in Disassoc of
STA with AP.  Fixed the issue by checking BT-AMP session is opened or
not. (btampContext->isBapSessionOpen). If its opened then only call
sme_RoamDisconnect.

637340: SAP does not beacon on Country code to which STA is associated
(CR 368474)

637497: completed tx statistic reporting functionality
(CR 370923)

637687: Support for MC/BC & hostARPOffload with ini to work instead of ioctl
Fixes MC frames are still passed to host driver although
McastBcastFilter=1, hostArpOffload=0 (CR 363707)

637850: Fix for SAP Auto Channel Selection feature
For SAP auto channel selection, WLAN host driver do scan to find less
noise channel. While performaing scan do not check for Active session
in CSR module and change CSR roam state to eCSR_ROAMING_STATE_SCANNING
state for scan command irrespective of session is active or not.  (CR
375829)

638022: Fix for CCX-STA to connect to CCKM-RSN AP
because of the incorrect if-else-if authentication type overwritten to
255 (CR 376930)

638142-638147: Fix compilation warnings caused by GCC -Wenum-compare
Recent versions of the GCC compiler include an "enum-compare" warning
which will emit a warning when an enum of one type is compared with an
enum of a different type.  The compiler has found occurrences of this
issue in the WLAN code.  Fix those issues such that enum comparisons
are restricted to enums of the same enumerated type.

638253: Implemented getAutoChannel ioctl for SoftAP
SoftAP does not have an ioctl to get auto channel. Implemented
getAutoChannel for SoftAP and fixed getchannel ioctl for softap (CR
369147)

638267: Fix GCC 4.6 compile failure in WDI due to inlining failure
WDI_GET_PAL_CTX() was defined to be an inline function, but the
function body was not available to the compilation units that refer to
it.  Other versions of the compiler have silently ignored this, but
GCC-4.6 is emitting an error "sorry, unimplemented: inlining failed in
call to 'WDI_GET_PAL_CTX': function body not available".  So we remove
the unnecessary "inline" directive from the definition of
WDI_GET_PAL_CTX().

638291: Fix WLAN suspended flag
If WLAN driver suspend fails becuase of a TX, RX or MC thread suspend
timeout, the suspended flag is not reset. So set the suspended flag
only after all three threads are suspended successfully.  (CR 375230)

638703: Disabling AMP flag
Disabling AMP flag from config file, so that we don't register with
HCI any more as BT AMP is not PORed (CR 375826)

Change-Id: I6234a721b9853c66c47949c685daa2b284d8eb4c