msm: acpuclock-krait: Remove use of QSB as a CPU or L2 clock source
QSB's rate is tied to the Apps fabric (or BIMC on 8974), which means
that its rate is unpredictable. When the CPU is running at a low
voltage, if the QSB clock source is selected, it's possible that the
CPU clock rate could increase beyond the safe limit for that voltage.

Instead of selecting QSB for power-collapse and hotplug scenarios,
select an always-on source with a predictable rate.

Change-Id: I301546fa8e652a444420043812b925a3deacbe00
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>
(cherry picked from commit c0055d20b1ec931286f21f208d967700a798403f)