msm: subsystem_restart: Don't use phase 3 SSR on the SGLTE
On the SGLTE platform, phase 2 SSR is required to address
the dependency between the internal and external modems. The
other subsystems (RIVA, LPASS etc) still need to do an
independent restart. In the absence of a matching restart
order, a request for a subsystem restart on a subsystem will
be treated as an  independent phase-3 like restart, even if
the current restart_level is phase 2. Therefore, don't allow
phase 3 on SGLTE platforms.

Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>
(cherry picked from commit e4ae75ea79620e3b59859d82aa1809a256e48364)

Change-Id: Ie7f9b8f966c1d77832486226e0f6d316717ec8e7
Signed-off-by: Sudhir Sharma <sudsha@codeaurora.org>