msm: video: wfd: Move domain specific mappings into subdevices
Move the domain specific mappings into subdevices to abstract away iommu
from wfd-ioctl.c.  This is especially needed as the encoder subdevices
for MSM8x60 and MSM8x74 differ in the iommu domain to which they map the
buffer to.  This target specific handling needs to taken care of by the
respective subdevices rather than by wfd-ioctl which should be target
agnostic.

Change-Id: I050d38842b2f6568d77b9e6419b109e2cfdacb76
Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>