msm: clock: Add clock mapping for rpm-regulator device
Add a CXO clock mapping for the rpm-regulator driver for all
targets which utilize a PMIC which requires CXO to be on in order
for its NLDOs to enable properly.  The workaround is only needed
if the target is utilizing a TCXO oscillator.

Change-Id: I7950641a5196e647d13de1e7529c97bd9894d8f0
Signed-off-by: David Collins <collinsd@codeaurora.org>
(cherry picked from commit a7d235346212f892758664179a3bfbb1d9350b9c)