gpio: msm: Clear the interrupt status before unmasking
After the change to keep the RAW_STATUS_EN always on, the
interrupt status bit was latched and could be set during
other (non-interrupt) activity on the GPIO line. This would
end up triggering the interrupt as soon as it was unmasked.
Hence, clear the interrupt status before unmasking the gpio
interrupt. This keeps the behavior and expectations same as
when the RAW_STATUS_EN was toggled and it prevented the
interrupt status to be updated while the interrupt was masked.

CRs-Fixed: 346861
Signed-off-by: Rohit Vaswani <rvaswani@codeaurora.org>
(cherry picked from commit 13f70c1bccde4bb91a331f14bab16d190681c162)

Change-Id: Ibcd55f01c7cb5c8b7255259abcf3240ae5b19a95
Signed-off-by: Sivasri Kumar Vanka <sivasri@codeaurora.org>