diag: Increase HSIC throughput by implementing multiple simultaneous reads
The HSIC implementation has been found to be too slow causing
the dropping of logs. The HSIC implementation is using a one-in-one-out
read buffer strategy. This strategy coupled with the high latency of
callbacks is leading to poor throughput, and thus the dropping of
logs. Modify the implementation to queue up multiple simultaneous
read requests on the HSIC. Manage these read buffers using a memory
pool.

Change-Id: I7cf334f54ebc41d0e222732e4cf68e158f015c67
CRs-Fixed: 373328
Signed-off-by: Dixon Peterson <dixonp@codeaurora.org>