Bluetooth: hci_ath: Configure GPIOs as part of bluetooth on.
Configuration of GPIOs should not be part of module init process
as gpios can be shared with other module.
Moved the configuration of gpios as part of driver open and releasing
them as part of driver close, so that there will not be any conflict
between different modules.

Change-Id: Iac507dd1e07e2cd4adb4c28058800fa4f51c9b52
Signed-off-by: Ram Mohan Korukonda <rkorukon@codeaurora.org>
(cherry picked from commit e359d865fd343481f86abeff5120d573493a72cc)