ASoC: WCD9310: Defer the setup of headphone removal detection
After the inserted accessory is detected as a valid headphone,
polling is performed on the microphone to correct the accessory
detection in case it was wrongly identified as a headphone.

It was observed that at the end of polling, headphone removal
interrupt was fired, causing headphone playback to stop. Fixed
by deferring the headphone removal setup until either polling is
complete or polling detects a removal

CRs-fixed: 384967
Signed-off-by: Bhalchandra Gajare <gajare@codeaurora.org>
(cherry picked from commit c1a831d55b209c0634eb94b091f32cab141e5e2f)

Change-Id: I4f54e42993de98eaac0f93662215a342c2ab0b90
Signed-off-by: Sivasri Kumar Vanka <sivasri@codeaurora.org>