msm: rpm-smd: Remove BUG if packet size is 0
When a interrupt for data event is received, it is possible that the SMD
packet could have been processed on other core executing with interrupts
locked. In this case, the data event would have scheduled a workqueue
which would result in 0 bytes for packet size. This is a valid scenario
and shouldn't throw a bug.

Signed-off-by: Mahesh Sivasubramanian <msivasub@codeaurora.org>
(cherry picked from commit 137cd700c1524ee274118d5dd6055d3673a2e8f7)

Change-Id: I90efeaa814ebec82542e46c35df754a1f380251c
Signed-off-by: Sudhir Sharma <sudsha@codeaurora.org>