msm: camera: Enable divert of RDI frames to userspace.
RDI frames from sensor may have to be processed depending
on the usecase. For eg: For a sensor capable of outputting
only YUV 422 data, the frame may have to be processed in the
post processing pipeline to convert it to the standard
YUV 420 format. Add support to divert these frames.

Change-Id: I5a0df274348ccf6cc2818eac7652ef83771d011b
Signed-off-by: Kiran Kumar H N <hurlisal@codeaurora.org>