arm: arch_timer: Enable timer in set_mode only
Currently, we enable the timer hardware and unmask its interrupt in
the set_next_event callback. However, during hotplug, this callback
can be called after the timer is already supposed to be disabled.

Since the timer will be completely turned back on by set_next_event,
we will soon receive an interrupt, potentially causing us to wake up
from hotplug.

Instead, continue unmasking the interrupt when set_next_event is
called, but only enable the timer hardware in the set_mode callback.
This means that, once arch_timer_disable is called, we are guaranteed
to not receive more timer interrupts until the cpu is intentionally
hotplugged back in.

Change-Id: I0e440f58a0ef6e453c7cfb0bbe2d7a657fbb041f
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>