msm: qdsp5: Fix for crash in mp3 driver
When resource allocation fails during mp3 playback, a crash
is observed. When such failure occurs, the driver tries to
access a handle that was never initialized. This is due
to missing flag check to determine whether there was any
memory allocation or not prior to failure. Added flag in error
cases to avoid such issues.

Change-Id: I6a772f4e1e61602c4b505c3d9b00fb4c7af939b0
CRs-Fixed: 394671
Signed-off-by: Chaithanya Krishna Bacharaju <chaithan@codeaurora.org>