ASoC: msm8974: LPA: Fix pause and next clip play issue
- When paused and press next button to play next song,
sometimes CMD_EOS fails to get Ack from LPASS and
is wait timeout for 5sec causing delay for next
playback start.
- In the failure case, even before trigger start
of driver is done,LPA driver receives pcm_close.
In this case, though EOS is issued, it is not
getting honored from LPASS.
- If trigger start not happen in LPA driver,
avoid CMD_EOS to LPASS as it will not be handled.

Change-Id: I3cb1f929c0e6c91f79c92949d10d0edf6c00081f
Signed-off-by: Phani Kumar Uppalapati <phanik@codeaurora.org>