msm: dcvs: force on the slack timer
For performance reasons the kernel should execute
a ONLINE call on TZ when the governor starts. The return
 values from ONLINE call are a new frequency and a new
 slack timer.

This new frequency would be queued up for the thread to set it up
on that core.

The way we handle the new slack timer is interesting. When the
governor starts we really dont know what the state of idle pulses
is, this is indicated by setting idle_entered = -1.  However we
want to start the slack timer. To handle this situation create
a function to force on the hrtimer without checking idle_entered or
pending_freq.

Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
(cherry picked from commit 973671aee4b5f2955a0bb85107022189bc12e75c)

Change-Id: I4cd20ad4fe3af8e980c47bca729aef524852dba3
Signed-off-by: Ram Kumar Chakravarthy Chebathini <rcheba@codeaurora.org>