power: pm8921-charger: Start unplug check worker for DC
It could happen that the device is powered up with a dc charger.
We should start the unplug check worker when the device starts with
dc charger or else we run the risk of not detecting the dc charger
removal.

(cherry picked from commit 8c0c84434d43e920867cfdae41ff50b72ba57843)

Change-Id: Ie5d7d95c68d084f19ff470dd3cf8e6682a985aa9
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>