msm: mdm: Add shutdown ioctl and send poweroff request to the mdm
When the whole phone is being powered off, a poweroff request needs
to be sent to the external modem so that it can shutdown gracefully.
This request needs to be triggered from userspace before kernel drivers
start unloading so that other drivers needed to send the request are
still available. The shutdown ioctl is provided for this purpose.
The request is sent over system monitor.

Change-Id: Ib63c7e5d822eb32e94a2fade10252c10f7eab46b
Crs-Fixed: 401598
Signed-off-by: Joel King <joelking@codeaurora.org>