msm: mdm: Add shutdown ioctl and send poweroff request to the mdm
When the whole phone is being powered off, a poweroff request needs
to be sent to the external modem so that it can shutdown gracefully.
This request needs to be triggered from userspace before kernel drivers
start unloading so that other drivers needed to send the request are
still available. The shutdown ioctl is provided for this purpose.
The request is sent over system monitor.

Change-Id: I84ab4d76caaddc30650c01ed010b3fedd90f7eaf
Crs-Fixed: 401598
Signed-off-by: Joel King <joelking@codeaurora.org>