USB: ci13xxx_udc: Apply posted writes workaround to control IN endpoint
The controller always does posted writes to transfer descriptors.
A transfer completion interrupt may be generated before dTD is
updated in system memory.  A software workaround is already in
place to give 10usec delay after receiving the completion interrupt
if the first dTD in the queue is still active.  But this delay
is given only for OUT endpoints.  In theory this problem is applicable
to both IN and OUT endpoints. But none of the function drivers indicate
any problem with IN endpoints.  Hence delay is not given for IN
endpoints.

The status phase of a control transfer may get stuck if the above
issue is hit for EP0 IN endpoint.  Hence apply the software workaround
for EP0 IN endpoint.

CRs-fixed: 409573
Change-Id: Ic07a4421c0066645459d2cc566f0df6691084a43
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
(cherry picked from commit 012800ec98ff32232b5fabe5eb8540a3152e65d0)

Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>