msm: kgsl: Turn off automatic hardware clock gating for A330
Turn off automatic hardware clock gating for A330 GPUs and
revert to the hardware default values. Turning the clock
gating on seems to push various parts of the hardware to the
edge without some mitigating patches on other subsystems. Going
back to the default values increases stability until all the
ducks are in a row.

Change-Id: Ic0dedbad8acea5d649332f18edf9dd6a88d9c164
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Carter Cooper <ccooper@codeaurora.org>