power: pm8921-bms: limit the number of tries for ocv calculation
When we want to force a certain percent charge we calculate the
expect percent charge (pc) from the expected RC (remaining charge
same as ocv based charge). We reverse lookup the ocv (open circuit
voltage) based on the pc. We double make sure that a forward lookup
matches the reverse lookup and if it does not within a percent limit
we repeatedly tweak the ocv by 5mV to get it to match.

This tweaking is unbounded and based on the battery profile could
result in a infinite loop. Limit the number of times we tweak the ocv
to prevent a possible watchdog bark.

CRs-Fixed: 406197
Change-Id: Iea874027a9e2f7ce326c8398f255e393fa7e4529
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>