power: pm8921-bms: handle restart without pmic power on
There are situation when the system undergoes a restart but the
pmic does not execute its power on sequence. In such case the OCV
value is not updated and the CC value is not cleared. Since OCV
and CC values came from the previous boot, we don't have to use
shutdown soc in this case. The use of shutdown soc without reset
of CC increase error of soc. So this patch remove use of shutdown
soc in restart case.

b/7327066

Change-Id: I2c5c5974a4c8ace076c1f8c56c856718b7140fb8