slim_msm: Retry initial sequence messages
Initial sequence messages (logical address assignment, and master
capability message) are very important messages to enable the
devices. It is possible that the device may NACK these messages
initially if it is in process of acquiring sync.
If these type of messages fail, retry a few times before actually
deeming the device unresponsive. Log the registers in case of NACKs
to better analyze the reason.

Change-Id: I89a42b20b53eae3a0a9d654dd70aa561bdcb5644
Signed-off-by: SathishKumar Mani <smani@codeaurora.org>