power: pm8921-bms: use pmic power reason to check restart
For accurate SoC calculation, bms need to know if target
was booted from reset or power down. Previously, minimum
CC value check was used to detect power on reason.
But this minimum CC value check is useless when charger is
connected. So, this patch use pmic power on reason to know
the exact power on reason.

Change-Id: I935261baf08f3c5df550c6c4eaa4648c5a37530e
Signed-off-by: Iliyan Malchev <malchev@google.com>