power: pm8921-charger: fix PC USB charging valid status
When PC USB charger is connected the driver ends up reporting
that is is not. This is because the driver incorrectly ends up
setting the target current. A non zero target is indicative
of a usb wall charger plugged in and the driver incorrectly
reports that the pc usb cable is unplugged.

(cherry picked from commit 728faaa0e60b27efa411d4413bd158f9118eadfb)

Change-Id: I9ec392ccf7c6f51ef83d2b075facea707ef3e08c
CRs-Fixed: 414741
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>