power: pm8921-charger: fix PC USB charging valid status
When PC USB charger is connected the driver ends up reporting
that is is not. This is because the driver incorrectly ends up
setting the target current. A non zero target is indicative
of a usb wall charger plugged in and the driver incorrectly
reports that the pc usb cable is unplugged.

(cherry picked from commit cda28f9ded515446e1de60531c4b71af080bdbe0)

Change-Id: I8b78f57cec1c0fb725540c4599e16231aae184c1
CRs-Fixed: 414741
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>