power: pm8921-charger: fix DC issue
The current code does not start the unplug check worker if dc
path is used. This is because it assumes an external charger
for the DC path.

Reorganize the DC insertion handler to check whether an external
charger is used. If not, call the unplug check worker since we
are running in a configuration where the DC path on pm8921 charger
is used.

Also update the code such that hkadc calibration is called when a
dc charger is plugged in, similar to when the usb charger is plugged
in. This is to avoid inaccuracies because of shifts in MBG when a
charger (either DC or USB) is inserted.

(cherry picked from commit 70ef382867b1aa6a0f2f62e2becd75c32b0de1df)

Change-Id: I396ce61eb74469e01aac1c00a26bf0dc3636f215
CRs-Fixed: 403147
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>