power: pm8921-charger: remove unused charging enable/disable api
This api could interfere with the internal working of the charger.
The end of charge detection is broken in hardware, the software workaround
relies on the enable/disable of the charging function. We don't want
this to be changed via some external api.

(cherry picked from commit d7cb1c7f49edbed2cc5f9943df6fc77cdc83ddf0)

Change-Id: I1d7ddaa1c4a1dbd0c579194bee306674f3d6cdea
CRs-Fixed: 417243
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>