power: pm8921-bms: handle restarts without pmic power on
There are situations when the system undergoes a restart but the
pmic does not execute its power on sequence. In such cases the
OCV value is not updated and the CC value is not cleared - they
carry values from the previous boot.

In such cases reset the CC value and measure the OCV. One can easily
detect the non power on case by checking the CC value. If it is above
5mAh, trigger this workaround. One can also use the power on module
register status to check if the pmic was restarted with a warm reset.

On similar lines we have seen issues when a very low OCV is recorded.
(5mV). This could be because the battery is not secure in the terminal
or a missing battery etc. Apply the same workaround for such cases.

(cherry picked from commit 3ae2d41848264e8099a9ce22015ee1779c906d64)

Change-Id: I5a0f82b79b04eeed7874efaf8dd60a36fb17da3c
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>