mfd: wcd9xxx: fix multi channel payload calculation
When multiple channels are used for a session (Rx/Tx), all bits in the
Rx/Tx port multichannel register needs to be Set. If a bit corresponding
to any channel is missing, the channel will not be aware that the other
channel (corresponding to the missing bit) is also part of the same group.
This will create ambiguity among the channels in the same PCM group and
will lead to shifting of Phase/delay in data in those channels.

(cherry picked from commit e4aaf089a885e32ab3b3428b6d1866599e006cb0)

Change-Id: I01daed81efb92eea43303a967c90c5b223c54339
CRs-fixed: 425803
Signed-off-by: Banajit Goswami <bgoswa@codeaurora.org>