msm: 8960ab: Add gfx3d axi clock and port to footswitch data
On 8960ab we need to assert and deassert the resets to the gfx3d
axi clock in addition to the core and interface clocks. Add the
clock to the list so that we reset the gfx core appropriately.

We also need to halt both bus ports, therefore add the second bus
port id.

CRs-Fixed: 428846
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
(cherry picked from commit 9071e51f27a02139a7ea56d839a1d00ca22fa345)

Change-Id: I5e7807f83cd0b299f25ba1200e9902849fedf70b
Signed-off-by: Vinod Kumar Pendyala <vpendy@codeaurora.org>