msm: 8960ab: Add gfx3d axi clock and port to footswitch data
On 8960ab we need to assert and deassert the resets to the gfx3d
axi clock in addition to the core and interface clocks. Add the
clock to the list so that we reset the gfx core appropriately.

We also need to halt both bus ports, therefore add the second bus
port id.

Change-Id: I05943906ed124658ab643d9bbce2bbca3f526a78
CRs-Fixed: 428846
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
(cherry picked from commit 9071e51f27a02139a7ea56d839a1d00ca22fa345)