power: pm8921-bms: set minimum average current for UUC
The current code accepts a very low current value when UUC is
calculated. This causes the UUC value to stay very low when the
phone is left idle. This is also the cause that prevents SOC
from going low as the temperature decreases.

Set the minimum average current to 250mA.

Change-Id: I472b9da10e9f45b3c254f8440526f3f9d5116931
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>