media: dvb: Fix un-mapping of external buffers
The output buffers of demux filters can be set in decoder
and in non-PES filters, it is separate from the video output
buffers. When filter is stopped the buffer should be un-mapped
always if it was mapped previously regardless of the filter's type.

Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>
(cherry picked from commit 2fe88e3f8caea60395f446b9e1716df57faeef80)

Change-Id: Ifb77aec0351483899dd0aa2582f5f22ba2462ed2
Signed-off-by: Sridhar Gujje <sgujje@codeaurora.org>