mmc: increase BKOPS polling duration before suspend
The timeout for the BKOPS polling is increased in order to allow the
card to complete the BKOPS before going to suspend.
In case we exit due to timeout we need to stop the BKOPS activity
performed by the card.

Signed-off-by: Maya Erez <merez@codeaurora.org>
(cherry picked from commit ea1385e6d7542d4e3ad510de94ead00bece7823e)

Change-Id: I6affa77960a2ec5bcd72cbbf7bf966c5b761ff9e
Signed-off-by: Maya Erez <merez@codeaurora.org>