msm: avs: Add API to restore the AVSCSR register
During power collapse, the AVSCSR register is always enabled and never
restored. This could result in AVS getting enabled on resume, when the
core is running at a frequency that doesn't support AVS.

Add new API, avs_enable, which restores the AVSCSR register based on the
saved AVSCSR register. Update the avs_disable API, to return the
AVSCSR register value prior to disabling the AVS.

Change-Id: Ide0334d5b78c7cb84df3631a7a6cbcc7393c7a51
Signed-off-by: Anji Jonnala <anjir@codeaurora.org>