msm: audio: qdsp6v2: Check return code from audio_aio_open()
When audio_aio_open() is triggered in blocking fashion, it  frees audio
resouces and returns error code. In the current implementation,  callers
did not check the error code which can lead to dereference to freed
resources and cause system crash.  Update the callers to check return
code and then return error back to application to avoid the issue. In
the meanwhile, update the audio_aio_open not to free memory for caller
and leave the caller to free the memory in error condition.

Change-Id: If3bc82f1704ff7cfa0814072eba871d8deb0fd54
CRs-fixed: 436653
Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>