CHROMIUM: ARM: tegra: add wake up event bindings for PMC
The PMC is the only device that can wake up the system from deep sleep
mode (i.e. LP0). There are some wake up events that in the PMC wake mask
registers can be used to trigger PMC to wake up the system. It defines
which devices or siganls can be the source to trigger PMC waking up. If
the devices support waking up system from deep sleep mode, then it needs
to describe a property for PMC wake up events. This patch defines the
properties for the usage.

BUG=none
TEST=the pmc wakeup source could be parsed in the DT

Change-Id: I9db75a54a923c789c6fac22cb5da146a93becf8c
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/176998
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>