UPSTREAM: ASoC: max98090: add digital mic mux to record path
The max98090 driver currently treats the digital mic enable as a supply
on the record path, causing the digital mic enable to always be turned on
when attempting to record.  This is incorrect, however, since the digital
mic enable is also a mux control where 0 selects the ADC output as input
to the record-path DSP and 1 selects the digital mic.  This patch adds
a virtual DMIC mux to the reocrd path so that we can switch between the
ADC and the digital mic for recording.

BUG=chrome-os-partner:18720
TEST=After audio support is added, observe that enabling the record
path does not force the digital mic enable bits on.

Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Signed-off-by: Mark Brown <broonie@opensource.wolfsonmicro.com>
(cherry picked from linux-next commit
	fd5f940f82cdc0132438a96559b27dd7fd574875)

Change-Id: I92b7a51747258e69822b81426fb9df9b8e7e48ed
Reviewed-on: https://gerrit.chromium.org/gerrit/57630
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/177661
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Commit-Queue: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>