msm_fb: Free iommu memory handles properly
Currently, the current memory handles are not sent to free list
unless there is a next subsequent play call on the pipe. After
we unset the pipes, unless there is a further overlay_play call
on those pipes, memory will be held. Change to increase one more
stage but free all the handles during unset.

Change-Id: I83e391348479a938dd3f986c7cfcfdc71c5bd3d8
Signed-off-by: Sakshi Sharma <sakshis@codeaurora.org>